Fix up content provider and hook
Now the content hook may run many times but the network tab
shows that Profile data is only being fetched once or twice
even as you navigate around the app, so basically the job
is done. It still seems like the sidebar is rerendering
too much but as long as it's no re-fetching the data it's
acceptable for now.