Make the profile/context process more reliable
Using lessons from
https://ruanmartinelli.com/posts/supabase-authentication-react
and moving the "fetch profile" step into the main
same useEffect in the GlobalStateProvider instead of useSWR.
Also no longer using useMemo for the final value because it
reliably only runs once, and again when auth state changes.

Authstate change clears or resets the user state hook which
is the one dependency on the useEffect that reloads the profile
and decks info.