[WIP] Add pages and forms to reset password
Add /forgot-password and /set-new-password,
send password reset emails and allow user
to choose a new password.

These don't work yet but the pages are there while I take
a break and read up on the docs for how to properly intercept
the user's verified link-click and pass it into the
set-new-password page.