Add more null checks and loading checks
This commit is a bit of work in the direction of
trying to make sure things aren't rendering before
context and profile have fully loaded, but it's not
working.

In some places, checking for !profile isn't enough;
we have to also check for !profile?.languages_spoken,
which is strange and as yet unexplained. In this case
when the content renders before context and profile
have loaded, it also doesn't update when that profile
information is updated. So this is the real crux of
the issue that will need to be solved at some point,
but for now it's not too too disruptive to the user
experience so I'm committing and moving on.