Add checkboxes for languages_spoken
In addition to the primary interface enhancement here,
(so that users don't have to edit an array by hand,)
this commit makes a bunch of (frankly somewhat arbitrary)
changes to both the profile index page and the way the
context object is handled, logged, etc.

The problems I was trying to solve are 2:
1. defaultValue only runs on first render; need to switch
to controlled inputs to solve this. The UI problem is only
when updating a primary language should check the
corresponding box in ther languages_spoken section, when
it's not already checked from before.
2. sometimes the context object updates but child components
don't re-render. This is really weird and I have to treat it
as a larger investigation and possible refactor.