Add a "shy" parameter for the sidebar
The "shy" parameter is for browsing pages where you want the
sidebar to be available but not open by default.

The sidebar code is sort of complex due to handling both open/closed
state and responsive styling. Now we've added Shyness. So, for future
reference, to perhaps refactor later, here are the pieces:

* on small screens the layout of the sidebar is: fixed position with
  a black overlay occluding the page below
* on medium and up screens the layout of the sidebar is: sticky and
  pinned to the left side of the screen
* shyness does not matter on small screens, but on medium-and-up
  screens it changes the behavior to resemble small-screen sidebars.
  i.e. all small screens are shy, but not all medium screens are shy
* on medium and up screens, non-shy/default sidebars are more flambouyant
  and will display the sidebar and hide the open/close button, regardless
  of open state
* on small screens and/or shy sidebars, the open/close button is
  always available, and controls visibility

Anyway the UX is pretty nice now. I would like to do something
about how you get a kind of double-menu with the sidebar and footer,
but I am not fussed about it now.