WIP: add react-query and switch to useDeck & useDecks
TBH I don't know why this isn't working. More to the point, the
console.log statements I added in 'use client' components appear
to be running on the server, and not the client. As one would
expect, the queries are also not running, so the response stays
in status==='loading'. (Actually this is an indicator that it's
kind of working? the hook is returning that status data.)