Completely remove urql and swr
I think SWR must have been out of use for ages, but now they are
both totally gone; all data fetching in the /app folder is running
through graphql-request in /app/data/fetchers, all using the same
queries in queries.js. All client requests are going through
react-query, using the same fetcher functions.

Supabase's js client is still handling sessions, running
getSession inside every fetch, which should be immediate and not a
problem(?) except when the token needs to be refreshed in which
case this async action is good for the user anyway!

Only the sidebar and the /profile pages load data from the old
.select('profile') approach provided by supabase-js.