Revert all .tsx to .jsx
This became more trouble than it is worth because useQuery does not
type the error messages / or convert them. So I'm having a lot of
trouble with the types and I don't know enough to fix them ✌️

Apparently the preferred approach is to use error boundaries and
pass the error boundary component to useQuery. This would be better
architecture/DX anyway so I will do this, but later.