Rework the BigModal to fetch its own data
This is a big commit with a lot of changes, but it now makes the BigCard
component into a bit of its own interface. It is still a WIP as I need
to implement the "edit status" button row for when a card is already
in your deck.

And somewhere along the way the "which options are disabled because I
already have this card in my deck?" array stopped working so I'll need
to fix that too.