Refactor addNewCard mutation, BigCard, etc
This mutation is a part of BigCard which is included in Browse, which
is included on the page that shows a deck and its different cards and
their states. To avoid adding new queries I ended up changing from a
fetcher that fetches all phrases in language, and one that fetches the
language itself, including all phrases and any deck associated (and
that passes RLS, i.e. belongs to the auth'd user).

So it's a kind of sprawling change but it simplifies a lot, and I think
it also is a kind of north star pointing to the silly bits, e.g.: maybe
the current deck's ID should be held in Context; this seems like it's
exactly what context is for.