Replace GlobalStateProvider with AuthProvider
This commit also renames useGlobalState to useAuthState, removes the last remnants
of the session key from the context object (because it carries the access token so
I just don't want it getting passed around any more than it has to). Also, since
we are now very clearly only using this AuthContext in login pages and user
management pages, we switched to the more expensive getUser which checks the current
session against the database, just to be sure everything is up-to-date. This is
probably unnecessary but it's a single HTTP request that may prevent some error
handling later on so why not 🤷.