Upgrade supabase-js to get that INITIAL_SESSION
The Initial State auth state change event allows us to write the
useAuthState hook in such a way that it only runs an effect once
to subscribe to supabase.auth.onAuthStateChange, and then the
callback function will run when the page first loads, so no separate
case has to be written to load auth state for the first time. This
is much smoother and simpler!