Simplify the auth provider
This commit simplifies the auth provider to rely on the callback for
onAuthStateChange so that no initial, special-cased getUser is required.
It also should prevent the useEffect in the Auth Provider from registering
duplicate listeners (perhaps I should be using useCallback for these), and
corrects the syntax for the listener.unsubscribe() for the effect's return/
unmounting behavior.