Fix a sort of race condition with the login/redirect
The LoginForm component was using the useAuth hook to see if the visitor
is logged in, and the useRouter hook to get a router and then do router.push
when detecting someone logged in. This seems to have created a race condition
where router.push gets called before the useRouter hook is done with its
whole setup (not quite sure the correct technical term, but the error message
was that the Router component is still rendering).

This router.push isn't needed on the modal version of the login form, so I
added it to a client component on the /login route itself, which wraps the
LoginForm. This new ClientPage only checks whether the user is auth'd, and
performs a redirect if so. And returns null, otherwise, it returns the login
form. Now the login form no longer checks to do this redirect in the beginning,
it only does it in the onSubmit/.success function. But it does still check
the isAuth status and does an early return null just to save clutter.