Consolidate most layouts into the root layout
This change moves the sidebar into the root layout and sets the utility
classes needed to make sure it plays nicely with the main content area:
the main content area will be centered both vertically and horizontally
within whatever space it has (either the full screen, or screen minus the
sidebar's width).

Now every page and layout share the root layout's "shy" sidebar, with no
variations based on my mood at the time of building a given screen.
With the Sidebar always available, the Footer is no longer needed. Other
pages/layouts have lost their complex mix of different ways to gett
content centered and aligned. Other layouts now can take a break; some
were retired, and others now only do the 1 thing they need to do, like
showing the avatar section or presenting the auth challenge to a section
of routes.

It is a significant improvement and code simplification, thanks to the
new NextJS layouts and a single decision not to offer a different
navigation UX to different parts of the app.