Refactor the add-a-new-card form
This commit adopts the new convention for colocating a Page, client Form,
and the post-something.ts mutation function. But I opted not to also
refactor the submit logic as well because it is already using mutations
(just, kind of maybe a bit backwards) and it's a complicated-enough function
with multiple promises being awaited... that it seems like a rabbit hole
I don't need to go down now.