Maybe fix a useContext error
There is an error happening on the change-email page where it seems
like we're trying to access the user object before it's ready, (the
actual error is trying to access property of useContext when it is
null; the AuthContext is the only context being used on that page).

I'm not sure why the previous code wouldn't also have been resilient
to a delayed response, or why other pages weren't breaking, but this
one would error almost every time I performed a hard refresh on it,
while others would not, and this one wouldn't error if navigating
from another page where the useAuth hook had been called. And it
seems to fix it.