Apply new rules for sizing the main app container
We now haave a consistently sized container for the app to take place in,
using container queries to make sure it works nicely whether the sidebar
is open or closed. This will be key for next steps of getting a sticky
nav bar that works across the site, but still lives inside these layouts,
without each layout having to re-do a bunch of work.

We are starting to adopt a paradigm where layouts are responsible for
setting their own max width, and whether they want to be `mx-auto` or not.
There is some repetition here but it is pretty minimal and makes sense,
(just set max width and center,) and if you leave it out, like on the
getting-started page, you still just have a nice container to work with.