Add a really basic and very restrictive/pro-privacy privacy policy
This is a little bit strange because the privacy policy gives an air
of completion and professionalism but it is just ripped from a free
privacy policy generator and then modified a bit to ensure it's not
talking about a Company and that it has a TL;DR that makes a few things
super clear: namely what data we are collecting, that your user name
and avatar are public, that we're not doing any other/hidden cookie-ing
or tracking, that we're not sharing or selling any data with anyone.

I thought it was better to do this now than wait _without any_ privacy
policy for an indefinite amount of time and then maybe have actual
users to worry about back-filling privacy consent for, and so on.
I will also have to add it to the login screen and the sidebar but
I think it's okay to do that in another PR. This one is about the home
page, and in this sense the existence of the privacy makes that home
page update complete and also goes a long way to fulfilling some basic
obligations, despite not having gone through the site and finished
making it discoverable to the level we would want. Soon.