Add a better Error component
The initial ErrorList component was meant to handle any-and-all error
situations so you could kind of just throw error objects at it and it
would kinda vaguely figure out what to do. But now we want to be a little
more intentional about such things, check for specific error codes, give
specific messages, destructure the error message intentionally in the
originating component and pass a simplified message to the component.

I'm not removing ErrorList right now but we should just kind of phase
it out as we go.