Replace more types with generated versions, and fix warnings
I keep having to list out all these types I wanted included and excluded
all over the place. It's kind of defeating the purpose of having the
types generated for me? Kind of? There is a certain ground-proofing going
on with the generated types; even if I'm still writing things by hand, I'm
doing it based on the knowledge they provide.