Stop using these Scalars
I think this was left over from the graphql days, but it is just adding
further abstraction and making code less readable, while most of the scalar
types are just aliases for "string" or "any".

This also adds a type "uuid" which is just an alias for "string", using lowercase
to look like a native/primitive type, even though it's not. (Not positive how I
feel about this. The database types use "string" for the uuids.