Switch from phrase_relation view to user_deck_plus.relations
!This could break some things!

This commit removes the phrase_relation view (it wasn't working w fkeys),
renames phrase_see_also to phrase_relation (always should have been the name),
and adds a view user_card_plus which contains an ARRAY column which lists all
the phrase_ids of related phrases (regardless of whether they are related from
the left or from the right).

Because our data model is keeping one store for deck (user) data and one store
for (public) language/phrase data, we don't need the database to have these
foreign-keys and automagically bring stuff in because the client is just going
to use javascript to fetch the phrase from the query cache, so this array of
related phrase IDs is giving us all the power the client actually wants, with
much less complexity.