Move prependAndDedupe to ts and rewrite it
Currently the function is only used in one place (raising the question
as to why it is in a lib file... but I digress) so I went to the page
in question (getting-started) and checked it out. the array-handling is
working just fine but I noticed there is this error message after the
form submits... which is really weird. so I made it go away. (but left
a TODO in the code, to find out... wth.)