Restructure pre-fetching w DeckPrefetch and DeckLoaded
This is so we can return the 'full' query with the same exact format
we will use in the query cache (DeckLoaded). (The only difference for
now is that the query cache uses ['deck', lang, 'card', pid] whereas
the preload query returns an array called 'cards'.

We might want to resolve this difference in the future, perhaps using
a single lookup object with keys (i.e. deck[pid] = card), but for now
we have the array of pids and we can loop that and use it to fetch
both phrase and card because they always have the same pid.