Change the name of the CardsMap to cards (from card)
The previous name was a remnant of the system where each card would
be cached individually and we were doing much more management of
the cache and very specific cache keys.

Then we were caching the response from the prefetch, and then splitting
it into the meta data, and the pids array, and caching them on their own
keys, and then caching each individual card on its own key.

But now we just have the one map object for the cards, and there's no
longer a `'deck', lang, 'card', pid` cache entry for each card so we
don't need the strange name any more!