Switch all the hooks to use this useLang approach
The "new" approach added the `const lang = useLang()` in to the base
hooks useDeckQuery and useLanguageQuery, so if they don't receive the
lang parameter from whatever is invoking them, they just grab it from
params and keep going. So now we don't need to pass a lang argument
as long as it's the same lang as the part of the route tree we're in,
so e.g. on any route like `deck/[lang]/some-path` we can go from:

```
const lang = deck.meta.lang
const pidsQuery = useDeckPids(lang)
```
to
```
const pidsQuery = useDeckPids()
```

Much easier! You can still pass the language to access a data-point
outside your current URL tree, like so: `useDeckPids(otherLang)`.
And thanks to the fact that this query is just a select of the main
useDeckQuery, it will go ahead and fetch the entire dataset for that
language to avoid having the app go into a totally different request-
making mode when accessing data outside of the current url tree.