Add these context-based ways of accessing the same data
So, in theory these should be the same, performance-wise, as the
hooks. They have a different return signature (only return data)
but as long as the hooks are used within the AppDataProvider, they
might perform identically. So I just have both of them in place for
now and I may test them both out to see how they compare.

The one difference here is you can't access things outside of the
current language context. Not without jumping ship and using the
hook instead. So perhaps the final implementation will be a bit more
like the current approach with the optional lang parameter, and when
lang is not set we simply use the pointer stored in context.

Based on: https://tkdodo.eu/blog/react-query-and-react-context