useLang instead of drilling lang={lang} all the time
This leaves a lot of the drilling in place because passing props
allows us to keep more components pure and server-renderable. The
new rule is to pass lang to any component that needs it, but not
further.

This also creates a new file lib/hooks -- where I'll probably move
some of the other hooks created in this process so I can drain out
this /api/ folder and eventually retire lib/data/hooks.