Update a couple things to use pids-style lists
In this commit we stop looping over whole arrays of card-phrases
and now we loop over arrays of phrase_ids and we access the needed
phrase or card from the map object. There are a couple of weirdnesses
popping up already. We want to easily access a list of phrase_ids for
just the active cards, or just the skipped cards or just the learned
cards. So for now we've just modified the old hook to provide this,
but soon we should make some filter function, or just pre-populate
those three lists along with the full one.

We've also migrated 4 files over to tsx! Typescript types aren't very
fleshed-out but everything passes.