Fix profile page not loading, with isPending
This was actually 2 problems... handling unloaded state with placeholder
data and then having weird things with async state + form field values
meant that it wasn't re-rendering when data came in. So we removed the
placeholder data and switched to isPending instead to avoid the initial
problem. I'll refactor that form another time